Refactor code for better organization

+++ src/MapClass/HolderClass/WeaponHolder.cs
@@ -9,15 +9,13 @@ namespace Dungeon.src.MapClass.HolderClass
     public class WeaponHolder(int x, int y, Weapon weapon) : Holder(x, y)
     {
         private Weapon weapon = weapon;
-
         public Weapon WeaponHold { get { return weapon; } set { weapon = value; } }
 
         public new void Draw(SpriteBatch spriteBatch)
         {
-            Rectangle weaponRectangle = new((int)hitbox.X + (hitbox.Width - weapon.Width) / 2, (int)hitbox.Y - weapon.Height - 10, weapon.Width, weapon.Height);
+            Rectangle weaponRectangle = new(hitbox.X + (hitbox.Width - weapon.Width) / 2, hitbox.Y - weapon.Height - 10, weapon.Width, weapon.Height);
             spriteBatch.FillRectangle(hitbox, Color.White);
             spriteBatch.FillRectangle(weaponRectangle, Color.Red);
-
         }
 
         public Weapon SwitchWeapon(Weapon weapon)
